+++ src/main/java/com/olbimacoojam/heaven/domain/Room.java
@@ -1,7 +1,6 @@
 package com.olbimacoojam.heaven.domain;
 
 import com.olbimacoojam.heaven.game.Game;
-import com.olbimacoojam.heaven.game.User;
 import lombok.Getter;
 
 import java.util.List;
