+++ src/main/java/com/olbimacoojam/heaven/service/RoomService.java
@@ -4,7 +4,7 @@ import com.olbimacoojam.heaven.domain.Room;
 import com.olbimacoojam.heaven.domain.RoomFactory;
 import com.olbimacoojam.heaven.domain.RoomRepository;
 import com.olbimacoojam.heaven.dto.RoomResponseDto;
-import com.olbimacoojam.heaven.game.User;
+import com.olbimacoojam.heaven.domain.User;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
